✨ add Ceros as an annotation with renderer and source
This supports a paste from the Ceros HTML snippet into an annotation and rendering back to the full HTML
