+++ packages/@atjson/offset-annotations/src/annotations/ceros-embed.ts
@@ -1,0 +1,10 @@
+import { ObjectAnnotation } from "@atjson/document";
+
+export class CerosEmbed extends ObjectAnnotation<{
+  url: string;
+  aspectRatio: number;
+  mobileAspectRatio?: number;
+}> {
+  static vendorPrefix = "offset";
+  static type = "ceros-embed";
+}
