+++ packages/@atjson/offset-annotations/src/annotations/index.ts
@@ -1,6 +1,7 @@
 export * from "./blockquote";
 export * from "./bold";
 export * from "./caption-source";
+export * from "./ceros-embed";
 export * from "./code";
 export * from "./facebook-embed";
 export * from "./fixed-indent";
