+++ packages/@atjson/renderer-html/test/renderer-test.ts
@@ -1,6 +1,7 @@
 import OffsetSource, {
   Blockquote,
   Bold,
+  CerosEmbed,
   Code,
   Heading,
   HorizontalRule,
@@ -355,5 +356,50 @@ describe("renderer-html", () => {
     });
 
     expect(Renderer.render(doc)).toEqual("<u>Hello</u>");
+  });
+
+  describe("ceros", () => {
+    test("without mobile aspect ratio", () => {
+      let doc = new OffsetSource({
+        content: "\uFFFC",
+        annotations: [
+          new CerosEmbed({
+            id: "test",
+            start: 0,
+            end: 1,
+            attributes: {
+              url: "//view.ceros.com/ceros-inspire/carousel-3",
+              aspectRatio: 2
+            }
+          })
+        ]
+      });
+
+      expect(Renderer.render(doc)).toMatchInlineSnapshot(
+        `"<div style=\\"position: relative;width: auto;padding: 0 0 50%;height: 0;top: 0;left: 0;bottom: 0;right: 0;margin: 0;border: 0 none\\" id=\\"experience-test\\" data-aspectRatio=\\"2\\"><iframe allowfullscreen src=\\"//view.ceros.com/ceros-inspire/carousel-3\\" style=\\"position: absolute;top: 0;left: 0;bottom: 0;right: 0;margin: 0;padding: 0;border: 0 none;height: 1px;width: 1px;min-height: 100%;min-width: 100%\\" frameborder=\\"0\\" class=\\"ceros-experience\\" scrolling=\\"no\\"></iframe></div><script type=\\"text/javascript\\" src=\\"//view.ceros.com/scroll-proxy.min.js\\" data-ceros-origin-domains=\\"view.ceros.com\\"></script>"`
+      );
+    });
+
+    test("with mobile aspect ratio", () => {
+      let doc = new OffsetSource({
+        content: "\uFFFC",
+        annotations: [
+          new CerosEmbed({
+            id: "test",
+            start: 0,
+            end: 1,
+            attributes: {
+              url: "//view.ceros.com/ceros-inspire/carousel-3",
+              aspectRatio: 2,
+              mobileAspectRatio: 3
+            }
+          })
+        ]
+      });
+
+      expect(Renderer.render(doc)).toMatchInlineSnapshot(
+        `"<div style=\\"position: relative;width: auto;padding: 0 0 50%;height: 0;top: 0;left: 0;bottom: 0;right: 0;margin: 0;border: 0 none\\" id=\\"experience-test\\" data-aspectRatio=\\"2\\" data-mobile-aspectRatio=\\"3\\"><iframe allowfullscreen src=\\"//view.ceros.com/ceros-inspire/carousel-3\\" style=\\"position: absolute;top: 0;left: 0;bottom: 0;right: 0;margin: 0;padding: 0;border: 0 none;height: 1px;width: 1px;min-height: 100%;min-width: 100%\\" frameborder=\\"0\\" class=\\"ceros-experience\\" scrolling=\\"no\\"></iframe></div><script type=\\"text/javascript\\" src=\\"//view.ceros.com/scroll-proxy.min.js\\" data-ceros-origin-domains=\\"view.ceros.com\\"></script>"`
+      );
+    });
   });
 });
